+++ app/src/main/java/com/mlsdev/mlsdevstore/presentaion/bottom_navigation/MainActivity.kt
@@ -3,21 +3,15 @@ package com.mlsdev.mlsdevstore.presentaion.bottom_navigation
 import android.os.Bundle
 import androidx.annotation.IdRes
 import androidx.databinding.DataBindingUtil
-import androidx.fragment.app.Fragment
-import androidx.fragment.app.FragmentTransaction
 import androidx.lifecycle.ViewModelProviders
+import androidx.navigation.Navigation
 import com.mlsdev.mlsdevstore.R
 import com.mlsdev.mlsdevstore.databinding.ActivityMainBinding
 import com.mlsdev.mlsdevstore.presentaion.BaseActivity
-import com.mlsdev.mlsdevstore.presentaion.account.AccountFragment
-import com.mlsdev.mlsdevstore.presentaion.cart.CartFragment
-import com.mlsdev.mlsdevstore.presentaion.store.StoreFragment
-import com.roughike.bottombar.OnTabReselectListener
-import com.roughike.bottombar.OnTabSelectListener
+import com.mlsdev.mlsdevstore.presentaion.utils.NavUtils
 
 class MainActivity : BaseActivity() {
     lateinit var binding: ActivityMainBinding
-    lateinit var navigationController: BottomNavigationController
     lateinit var viewModel: BottomNavigationViewModel
 
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -43,33 +37,7 @@ class MainActivity : BaseActivity() {
     }
 
     private fun initBottomNavigation() {
-        navigationController = BottomNavigationController()
-        binding.bottomBar.setOnTabSelectListener(navigationController)
-        binding.bottomBar.setOnTabReselectListener(navigationController)
-    }
-
-    inner class BottomNavigationController : OnTabSelectListener, OnTabReselectListener {
-
-        override fun onTabReSelected(tabId: Int) {
-
-        }
-
-        override fun onTabSelected(tabId: Int) {
-            val fragment: Fragment = when (tabId) {
-                R.id.navigation_item_cart -> CartFragment()
-                R.id.navigation_item_account -> AccountFragment()
-                else -> StoreFragment()
-            }
-
-            replaceFragment(fragment)
-        }
-
-        private fun replaceFragment(fragment: Fragment) {
-            supportFragmentManager.beginTransaction()
-                    .setTransition(FragmentTransaction.TRANSIT_FRAGMENT_FADE)
-                    .replace(R.id.content, fragment)
-                    .commit()
-        }
-
+        val navController = Navigation.findNavController(this, R.id.nav_host)
+        NavUtils.setupWithNavController(binding.bottomBar, navController)
     }
 }